+++ src/services/erp/automation/assigment-rule/enum.js
@@ -1,0 +1,19 @@
+export const SHIFTS = {
+  FIRST_SHIFT: "Ca1",
+  SECOND_SHIFT: "Ca2"
+};
+
+export const ASSIGNMENT_RULES = {
+  HN: {
+    "name": "Lead-Assignment-Sales-HN",
+    "region_name": "CRM-REGION-SOURCE-0000001"
+  },
+  HCM: {
+    "name": "Lead-Assignment-Sales-HCM",
+    "region_name": "CRM-REGION-SOURCE-0000002"
+  },
+  CT: {
+    "name": "Lead-Assignment-Sales-CT",
+    "region_name": "CRM-REGION-SOURCE-0000003"
+  }
+};
