+++ src/services/erp/automation/index.js
@@ -1,0 +1,5 @@
+import AssignmentRuleService from "./assigment-rule/assigment-rule";
+
+export default {
+  AssignmentRuleService: AssignmentRuleService
+};
