+++ src/services/erp/index.js
@@ -2,10 +2,12 @@ import OrderService from "./order";
 import AddressService from "./address";
 import ContactService from "./contact";
 import CustomerService from "./customer";
+import Automation from "./automation";
 
 export default {
   OrderService: OrderService,
   AddressService: AddressService,
   ContactService: ContactService,
-  CustomerService: CustomerService
+  CustomerService: CustomerService,
+  Automation: Automation
 };
